+++ gwcs/extension.py
@@ -1,4 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
+import importlib.resources
+
 from asdf.extension import ManifestExtension
 from .converters.wcs import (
     CelestialFrameConverter, CompositeFrameConverter, FrameConverter,
@@ -40,10 +42,12 @@ WCS_MODEL_CONVERTERS = [
 # The order here is important; asdf will prefer to use extensions
 # that occur earlier in the list.
 WCS_MANIFEST_URIS = [
-    "asdf://asdf-format.org/astronomy/gwcs/manifests/gwcs-1.0.0",
+    f"asdf://asdf-format.org/astronomy/gwcs/manifests/{path.stem}"
+    for path in sorted((importlib.resources.files("asdf_wcs_schemas.resources") / "manifests").iterdir(), reverse=True)
 ]
 
-
+# 1.0.0 contains multiple versions of the same tag, a bug fixed in
+# 1.0.1 so only register 1.0.0 if it's the only available manifest
 TRANSFORM_EXTENSIONS = [
     ManifestExtension.from_uri(
         uri,
@@ -51,7 +55,9 @@ TRANSFORM_EXTENSIONS = [
         converters=WCS_MODEL_CONVERTERS,
     )
     for uri in WCS_MANIFEST_URIS
+    if len(WCS_MANIFEST_URIS) == 1 or '1.0.0' not in uri
 ]
+
 
 def get_extensions():
     """